requirements.txt
github-utils-api
atlassian-python-api